+++ kyaml/filesys/fileinfo.go
