+++ kyaml/filesys/util_test.go
@@ -1,11 +1,12 @@
-package filesys_test
+// Copyright 2021 The Kubernetes Authors.
+// SPDX-License-Identifier: Apache-2.0
+
+package filesys
 
 import (
 	"os"
 	"path/filepath"
 	"testing"
-
-	. "sigs.k8s.io/kustomize/api/filesys"
 )
 
 // Confirm behavior of filepath.Match
